requirements.txt
Flask
tensorflow
numpy
Pillow
Flask-Cors
gunicorn  # Important for production web servers